requirements.txt
asgiref==3.8.1
Django==5.0.4
djangorestframework==3.15.1
exceptiongroup==1.2.0
factory-boy==3.3.0
Faker==24.9.0
inflection==0.5.1
iniconfig==2.0.0
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==24.0
pluggy==1.4.0
pytest==8.1.1
pytest-django==4.8.0
pytest-factoryboy==2.7.0
pytest-html==4.1.1
pytest-metadata==3.1.1
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.11.0
